refactor: split extended-types into separate files
Move MulmoBeat/MulmoScript related schemas to mulmoBeat.ts
in preparation for adding viewer data extension types.
